+++ src/components/FormInput/FormWrapper.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { InputContainer, Label } from './style';
+
+const FormWrapper = ({ label, id, children }) => {
+  return (
+    <InputContainer>
+      <Label htmlFor={id}>{label}</Label>
+      {children}
+    </InputContainer>
+  );
+};
+
+export default FormWrapper;
