+++ src/components/FormInput/index.js
@@ -1,0 +1,5 @@
+import DefaultInput from './DefaultInput';
+import ImageFileInput from './ImageFileInput';
+import TextareaInput from './TextareaInput';
+
+export { DefaultInput, ImageFileInput, TextareaInput };
